requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
pandas>=1.3.0
matplotlib>=3.4.3
scikit-learn>=0.24.2
tqdm>=4.62.2
pillow>=8.3.1
opencv-python>=4.5.3
albumentations>=1.0.3
tensorboard>=2.6.0
seaborn